backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
langchain-community
python-dotenv
pydantic
chromadb
pymongo
tiktoken